docs(get): document tenant and task lookup helpers #151 #152
