+++ c8volt/tenant/convert.go
@@ -8,6 +8,7 @@ import (
 	"github.com/grafvonb/c8volt/toolx"
 )
 
+// fromDomainTenant converts the internal tenant model to the public facade payload.
 func fromDomainTenant(x d.Tenant) Tenant {
 	return Tenant{
 		TenantId:    x.TenantId,
@@ -16,6 +17,7 @@ func fromDomainTenant(x d.Tenant) Tenant {
 	}
 }
 
+// fromDomainTenants wraps converted tenant items with the total used by command and JSON views.
 func fromDomainTenants(xs []d.Tenant) Tenants {
 	items := toolx.MapSlice(xs, fromDomainTenant)
 	return Tenants{
@@ -24,6 +26,7 @@ func fromDomainTenants(xs []d.Tenant) Tenants {
 	}
 }
 
+// toDomainTenantFilter converts facade filter input into the shared tenant service filter.
 func toDomainTenantFilter(x TenantFilter) d.TenantFilter {
 	return d.TenantFilter{NameContains: x.NameContains}
 }
