+++ cmd/get_tenant_test.go
@@ -23,6 +23,7 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
+// Verifies tenant list output renders stable compact rows with count and sort order.
 func TestGetTenantListOutput_SortsAndRendersCompactRows(t *testing.T) {
 	resetTenantRenderFlags(t)
 	cmd := newTenantListTestCommand()
@@ -48,6 +49,7 @@ func TestGetTenantListOutput_SortsAndRendersCompactRows(t *testing.T) {
 	require.Less(t, strings.Index(output, "tenant-a"), strings.Index(output, "tenant-b"))
 }
 
+// Verifies tenant list columns remain aligned when IDs, names, and descriptions vary in length.
 func TestGetTenantListOutput_AlignsTenantIDAndNameColumns(t *testing.T) {
 	resetTenantRenderFlags(t)
 	cmd := newTenantListTestCommand()
@@ -74,6 +76,7 @@ func TestGetTenantListOutput_AlignsTenantIDAndNameColumns(t *testing.T) {
 	}, "\n"), output)
 }
 
+// Verifies keys-only tenant list output emits tenant IDs for pipeline use.
 func TestGetTenantListOutput_KeysOnlyUsesTenantID(t *testing.T) {
 	resetTenantRenderFlags(t)
 	flagViewKeysOnly = true
@@ -94,6 +97,7 @@ func TestGetTenantListOutput_KeysOnlyUsesTenantID(t *testing.T) {
 	require.Equal(t, "tenant-a\ntenant-b\n", output)
 }
 
+// Verifies JSON tenant list output uses the public tenant payload and shared success envelope.
 func TestGetTenantListOutput_JSONUsesTenantPayload(t *testing.T) {
 	resetTenantRenderFlags(t)
 	flagViewAsJson = true
@@ -133,6 +137,7 @@ func TestGetTenantListOutput_JSONUsesTenantPayload(t *testing.T) {
 	require.NotContains(t, output, "members")
 }
 
+// Verifies keyed tenant lookup renders a single tenant without list summary text.
 func TestGetTenantByKeyOutput_RendersSingleTenant(t *testing.T) {
 	resetTenantRenderFlags(t)
 	cmd := newTenantListTestCommand()
@@ -150,6 +155,7 @@ func TestGetTenantByKeyOutput_RendersSingleTenant(t *testing.T) {
 	require.NotContains(t, output, "found:")
 }
 
+// Verifies --filter is forwarded as a literal name-or-ID contains filter and renders matching tenants.
 func TestGetTenantListOutput_FilterPassesTenantIDOrNameContainsAndRendersMatches(t *testing.T) {
 	resetTenantRenderFlags(t)
 	flagGetTenantFilter = "dev"
@@ -175,6 +181,7 @@ func TestGetTenantListOutput_FilterPassesTenantIDOrNameContainsAndRendersMatches
 	require.Contains(t, output, "found: 2")
 }
 
+// Verifies filtered tenant searches return an empty count without the unfiltered access warning.
 func TestGetTenantListOutput_FilterEmptyResults(t *testing.T) {
 	resetTenantRenderFlags(t)
 	flagGetTenantFilter = "missing"
@@ -193,6 +200,7 @@ func TestGetTenantListOutput_FilterEmptyResults(t *testing.T) {
 	require.NotContains(t, logBuf.String(), "tenant search returned no tenants")
 }
 
+// Verifies empty unfiltered tenant searches warn that the configured client may lack tenant visibility.
 func TestGetTenantListOutput_UnfilteredEmptyResultsWarnsAboutTenantAccess(t *testing.T) {
 	resetTenantRenderFlags(t)
 	cmd := newTenantListTestCommand()
@@ -213,6 +221,7 @@ func TestGetTenantListOutput_UnfilteredEmptyResultsWarnsAboutTenantAccess(t *tes
 	require.Contains(t, logBuf.String(), "configured client may not have access to tenant resources")
 }
 
+// Verifies tenant filters are literal contains checks rather than regular expressions.
 func TestGetTenantListOutput_FilterKeepsPatternTextLiteral(t *testing.T) {
 	resetTenantRenderFlags(t)
 	flagGetTenantFilter = ".*"
@@ -235,6 +244,7 @@ func TestGetTenantListOutput_FilterKeepsPatternTextLiteral(t *testing.T) {
 	require.Contains(t, output, "found: 1")
 }
 
+// Verifies keyed tenant lookup supports keys-only output for command composition.
 func TestGetTenantByKeyOutput_KeysOnlyUsesTenantID(t *testing.T) {
 	resetTenantRenderFlags(t)
 	flagViewKeysOnly = true
@@ -250,6 +260,7 @@ func TestGetTenantByKeyOutput_KeysOnlyUsesTenantID(t *testing.T) {
 	require.Equal(t, "tenant-a\n", output)
 }
 
+// Verifies keyed tenant lookup JSON uses the public tenant payload and shared success envelope.
 func TestGetTenantByKeyOutput_JSONUsesTenantPayload(t *testing.T) {
 	resetTenantRenderFlags(t)
 	flagViewAsJson = true
@@ -278,6 +289,7 @@ func TestGetTenantByKeyOutput_JSONUsesTenantPayload(t *testing.T) {
 	require.NotContains(t, output, "members")
 }
 
+// Verifies tenant list reports the explicit unsupported-version failure on Camunda 8.7.
 func TestGetTenantCommand_V87ListReportsUnsupported(t *testing.T) {
 	cfgPath := writeRawTestConfig(t, `app:
   camunda_version: 8.7
@@ -297,6 +309,7 @@ apis:
 	require.Contains(t, string(output), "tenant search requires Camunda 8.8 or newer")
 }
 
+// Verifies keyed tenant lookup reports the explicit unsupported-version failure on Camunda 8.7.
 func TestGetTenantCommand_V87KeyedReportsUnsupported(t *testing.T) {
 	cfgPath := writeRawTestConfig(t, `app:
   camunda_version: 8.7
@@ -316,6 +329,7 @@ apis:
 	require.Contains(t, string(output), "tenant lookup requires Camunda 8.8 or newer")
 }
 
+// Verifies --key rejects whitespace-only values before any tenant lookup can run.
 func TestGetTenantCommand_RejectsWhitespaceKey(t *testing.T) {
 	resetTenantRenderFlags(t)
 	flagGetTenantKey = "   "
@@ -329,6 +343,7 @@ func TestGetTenantCommand_RejectsWhitespaceKey(t *testing.T) {
 	require.Contains(t, err.Error(), "tenant lookup requires a non-empty --key")
 }
 
+// Verifies tenant keyed lookup and literal filtering stay mutually exclusive.
 func TestGetTenantCommand_RejectsKeyPlusFilter(t *testing.T) {
 	resetTenantRenderFlags(t)
 	flagGetTenantKey = "tenant-a"
@@ -345,6 +360,7 @@ func TestGetTenantCommand_RejectsKeyPlusFilter(t *testing.T) {
 	require.Contains(t, err.Error(), "--key cannot be combined with --filter")
 }
 
+// Verifies tenant help documents list, keyed lookup, filtering, and JSON examples.
 func TestGetTenantHelp_DocumentsListKeyFilterAndJSONExamples(t *testing.T) {
 	output := executeRootForTest(t, "get", "tenant", "--help")
 
@@ -355,20 +371,24 @@ func TestGetTenantHelp_DocumentsListKeyFilterAndJSONExamples(t *testing.T) {
 	require.Contains(t, output, "./c8volt get tenant --key <tenant-id> --json")
 }
 
+// tenantCommandAPI lets command tests replace only the tenant facade methods under test.
 type tenantCommandAPI struct {
 	c8volt.API
 	searchTenants func(context.Context, tenant.TenantFilter, ...foptions.FacadeOption) (tenant.Tenants, error)
 	getTenant     func(context.Context, string, ...foptions.FacadeOption) (tenant.Tenant, error)
 }
 
+// SearchTenants delegates to the test-provided search function.
 func (a tenantCommandAPI) SearchTenants(ctx context.Context, filter tenant.TenantFilter, opts ...foptions.FacadeOption) (tenant.Tenants, error) {
 	return a.searchTenants(ctx, filter, opts...)
 }
 
+// GetTenant delegates to the test-provided keyed lookup function.
 func (a tenantCommandAPI) GetTenant(ctx context.Context, tenantID string, opts ...foptions.FacadeOption) (tenant.Tenant, error) {
 	return a.getTenant(ctx, tenantID, opts...)
 }
 
+// newTenantListTestCommand creates a tenant subcommand with captured output for direct view/helper tests.
 func newTenantListTestCommand() *cobra.Command {
 	cmd := &cobra.Command{Use: "tenant"}
 	buf := &bytes.Buffer{}
@@ -379,6 +399,7 @@ func newTenantListTestCommand() *cobra.Command {
 	return cmd
 }
 
+// resetTenantRenderFlags isolates global render and tenant selector flags between tenant command tests.
 func resetTenantRenderFlags(t *testing.T) {
 	t.Helper()
 	prevJSON := flagViewAsJson
@@ -400,10 +421,12 @@ func resetTenantRenderFlags(t *testing.T) {
 	flagGetTenantFilter = ""
 }
 
+// tenantTestLogger returns a discard logger for command helper tests that do not assert log output.
 func tenantTestLogger() *slog.Logger {
 	return slog.New(slog.NewTextHandler(io.Discard, nil))
 }
 
+// TestGetTenantCommand_V87ListReportsUnsupportedHelper drives list execution in a helper process to preserve exit behavior.
 func TestGetTenantCommand_V87ListReportsUnsupportedHelper(t *testing.T) {
 	if os.Getenv("GO_WANT_HELPER_PROCESS") != "1" {
 		return
@@ -416,6 +439,7 @@ func TestGetTenantCommand_V87ListReportsUnsupportedHelper(t *testing.T) {
 	Execute()
 }
 
+// TestGetTenantCommand_V87KeyedReportsUnsupportedHelper drives keyed execution in a helper process to preserve exit behavior.
 func TestGetTenantCommand_V87KeyedReportsUnsupportedHelper(t *testing.T) {
 	if os.Getenv("GO_WANT_HELPER_PROCESS") != "1" {
 		return
