+++ internal/services/tenant/v89/convert.go
@@ -9,6 +9,7 @@ import (
 	"github.com/grafvonb/c8volt/toolx"
 )
 
+// fromTenantResult converts the generated v8.9 tenant payload into the shared domain model.
 func fromTenantResult(x camundav89.TenantResult) d.Tenant {
 	return d.Tenant{
 		TenantId:    string(x.TenantId),
